Fixed one bug

+++ include/parser.h
@@ -76,7 +76,7 @@ private:
 private:
     std::stack<std::shared_ptr<SymbolTable>> _scopes;
     std::vector<Token> _tokens;
-    int64_t _position;
+    size_t _position;
     bool _failed;
 };
